Reorganised admin panel components
