+++ src/pages/academy/adminPanel/subcomponents/assessmentConfigPanel/DeleteRowCell.tsx
