+++ src/pages/academy/adminPanel/subcomponents/assessmentConfigPanel/NumericCell.tsx
