+++ src/pages/academy/adminPanel/subcomponents/userConfigPanel/UserConfigPanel.tsx
